Hoist createSettingStore into shared tests/helpers.ts
SonarCloud flagged 11.3% duplication on the previous commit because
the Map-backed createStore() helper was a near-copy of the one already
present in home-api.test.ts.

Move it into tests/helpers.ts as createSettingStore() and import it
from both test files. Net: zero behaviour change, single source of
truth, lint/typecheck/format/coverage all still clean, 397/397 tests.
